+++ src/main/java/com/gdsc/plogger/trashcan/data/dto/res/TrashcanGetRes.java
@@ -1,13 +1,14 @@
 package com.gdsc.plogger.trashcan.data.dto.res;
 
 import com.gdsc.plogger.trashcan.data.Trashcan;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class TrashcanGetRes {
     private Long id;
-    private double latitude;
-    private double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
     public TrashcanGetRes(Trashcan trashcan) {
         this.id = trashcan.getId();
